Detect other SMSSecure clients and offer to secure the session
- Replaced the AutoInitiateActivity with a simple alert dialog
- Stripped the whitespace tag off incoming as well as outgoing messages
  for display purposes
- Refactored the helper methods in the AutoInititiateActivity and Tag
  classes into a single class
