--- res/layout/auto_initiate_activity.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:paddingLeft="16dip"
-    android:paddingRight="16dip">
-
-    <TextView android:id="@+id/description_text"
-              android:layout_width="fill_parent"
-              android:layout_height="wrap_content"
-              android:layout_marginTop="5dip"
-              android:layout_marginBottom="5dip"
-              android:textAppearance="?android:attr/textAppearanceMedium"                
-              android:text=""/>
-
-    <LinearLayout android:layout_width="fill_parent"
-                  android:layout_height="wrap_content"
-                  android:orientation="horizontal"
-                  android:layout_marginTop="16dip"
-                  android:layout_marginBottom="16dip"
-                  android:gravity="right">
-
-        <TableLayout android:layout_width="wrap_content"
-                     android:layout_height="wrap_content"
-                     android:stretchColumns="*">
-            <TableRow>
-		        <Button android:id="@+id/initiate_button"
-		                android:layout_width="wrap_content"
-		                android:layout_height="wrap_content"
-		                android:text=""
-		                android:gravity="center"/>
-		
-		        <Button android:id="@+id/cancel_button"
-		                android:text="@android:string/cancel"
-		                android:layout_width="wrap_content"
-		                android:layout_height="wrap_content"
-		                android:gravity="center"/>
-            </TableRow>
-        </TableLayout>
-    </LinearLayout>
-</LinearLayout>
